Fix issues with normals
In this .obj file, each vertex is linked to one or several normals so that we can have both smooth and flat shading. I didn't take this into account. This is now fixed by duplicating all vertices as many time as they belong to faces. This is not an optimized solution because some vertices are duplicated for nothing. This solution is fine for now.