Added error when trying to access an out of bound position
And fixed the mystic issue with the h6 position (I couldn't find why, but it was only after a move to the h6 position that access to an out of bound position of the "allowedNextMove" array resulted in crashing...). I fixed it by checking that I wasn't accessing an out of bound position.